+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2022-05-11
+
+### Added
+
+-   CSRF token (linking a device / payout)
+-   show (a device) function
+-   hide (a device) function
+-   removeAutoRedeem function
+-   usage function
+-   devicesStatus function
+
+### Changed
+
+-   appVersions function is renamed to downloads() according to the new endpoint name
+-   rename function endpoint
+-   more customization for registerDevice function
+-   dependency update
+
 ## [1.3.1] - 2021-21-16
 
 ### Changed